refactor: Use React components
