--- cv-application/src/components/Experience.jsx
@@ -1,38 +1,0 @@
-/* eslint-disable react/prop-types */
-import '../styles/App.css'
-import { TextInput, TextareaInput } from './Inputs'
-
-export default function Experience ({ content }) {
-    return (
-        <div>
-            {content.map((item) => {
-                return (
-                    <form key={item.key}>
-                        <div>
-                            <label htmlFor="name">Company/Organization</label>
-                            <TextInput name="name" defaultValue={item.name}></TextInput>
-                        </div>
-                        <div>
-                            <label htmlFor="title">Position Title</label>
-                            <TextInput name="title" defaultValue={item.position}></TextInput>
-                        </div>
-                        <div>
-                            <label htmlFor="startDate">Start Date</label>
-                            <TextInput name="startDate" defaultValue={item.startDate}></TextInput>
-                        </div>
-                        <div>
-                            <label htmlFor="endDate">End Date</label>
-                            <TextInput name="endDate" defaultValue={item.endDate}></TextInput>
-                        </div>
-                        <div>
-                            <label htmlFor="description">Description</label>
-                            <TextareaInput name="description" defaultValue={item.description}></TextareaInput>
-                        </div>
-                        <button className="deleteButton">Delete Experience</button>
-                    </form>
-                );
-            })}
-        </div>
-        
-    );
-}
